+++ src/counted_float/_core/counting/__init__.py
@@ -1,3 +1,4 @@
 from ._builtin_data import BuiltInData
 from ._context_managers import FlopCountingContext, PauseFlopCounting
 from ._counted_float import CountedFloat
+from .verbosity import Verbosity